+++ src/api/api-utils.ts
@@ -1,2 +1,2 @@
 // when testing on expo choose computer's ip instead of localhost
-const baseUrl = "http://localhost:8080";
+export const baseUrl = "http://localhost:8080";
