+++ src/api/types.ts
@@ -1,0 +1,6 @@
+export interface Organization {
+  id: number;
+  imageUrl: string;
+  name: string;
+  isSubscribed: boolean;
+}
